+++ method-analyzer-core/src/test/java/com/synopsys/method/analyzer/test/core/report/ReportGeneratorTest.java
@@ -28,7 +28,9 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -46,6 +48,8 @@ import com.google.common.collect.Multimap;
 import com.google.gson.Gson;
 import com.synopsys.method.analyzer.core.model.MethodUse;
 import com.synopsys.method.analyzer.core.model.ReferencedMethod;
+import com.synopsys.method.analyzer.core.report.BrokenFileJson;
+import com.synopsys.method.analyzer.core.report.BrokenFilesReportJson;
 import com.synopsys.method.analyzer.core.report.MetaDataReportJson;
 import com.synopsys.method.analyzer.core.report.MethodIdJson;
 import com.synopsys.method.analyzer.core.report.MethodIdsReportJson;
@@ -69,7 +73,8 @@ public class ReportGeneratorTest {
     }
 
     @Test
-    public void simpleReport() throws Exception {
+    @SuppressWarnings("deprecation")
+    public void simpleReportLegacyNoBrokenFiles() throws Exception {
         Multimap<ReferencedMethod, MethodUse> references = HashMultimap.create();
         references.put(new ReferencedMethod("methodOwner", "methodName", Collections.singletonList("input"), "output"), new MethodUse("use", 1));
         references.put(new ReferencedMethod("methodOwner2", "methodName2", Collections.singletonList("input2"), "output2"), new MethodUse("use", null));
@@ -149,15 +154,116 @@ public class ReportGeneratorTest {
     }
 
     @Test
+    public void simpleReport() throws Exception {
+        Multimap<ReferencedMethod, MethodUse> references = HashMultimap.create();
+        references.put(new ReferencedMethod("methodOwner", "methodName", Collections.singletonList("input"), "output"), new MethodUse("use", 1));
+        references.put(new ReferencedMethod("methodOwner2", "methodName2", Collections.singletonList("input2"), "output2"), new MethodUse("use", null));
+
+        Map<Path, String> brokenFiles = new HashMap<>();
+        brokenFiles.put(Paths.get("broken/BFile.class"), "Caused by TestException at MethodUseAnalyzer.java:123");
+
+        Path result = reportGenerator.generateReport(references, brokenFiles, testReportDirectory, "simpleReport");
+
+        Assert.assertNotNull(result);
+        Assert.assertTrue(Files.exists(result));
+
+        Path resultExpandedDirectory = unzip(result);
+
+        Path expectedMetaData = resultExpandedDirectory.resolve("metaData.json");
+        Path expectedReferencedMethod = resultExpandedDirectory.resolve("referenced-methods").resolve("referenced-methods-0.json");
+        Path expectedReferencedMethodUses = resultExpandedDirectory.resolve("referenced-method-uses").resolve("referenced-method-uses-0.json");
+        Path expectedBrokenFiles = resultExpandedDirectory.resolve("broken-files").resolve("broken-files-0.json");
+
+        Assert.assertTrue(Files.exists(expectedReferencedMethod));
+        Assert.assertTrue(Files.exists(expectedReferencedMethodUses));
+        Assert.assertTrue(Files.exists(expectedBrokenFiles));
+
+        MetaDataReportJson metaDataReport = null;
+        MethodIdsReportJson methodIdsReport = null;
+        MethodReferencesReportJson methodReferencesReport = null;
+        BrokenFilesReportJson brokenFilesReport = null;
+
+        try (BufferedReader reader = Files.newBufferedReader(expectedMetaData)) {
+            metaDataReport = GSON.fromJson(reader, MetaDataReportJson.class);
+        }
+
+        try (BufferedReader reader = Files.newBufferedReader(expectedReferencedMethod)) {
+            methodIdsReport = GSON.fromJson(reader, MethodIdsReportJson.class);
+        }
+
+        try (BufferedReader reader = Files.newBufferedReader(expectedReferencedMethodUses)) {
+            methodReferencesReport = GSON.fromJson(reader, MethodReferencesReportJson.class);
+        }
+
+        try (BufferedReader reader = Files.newBufferedReader(expectedBrokenFiles)) {
+            brokenFilesReport = GSON.fromJson(reader, BrokenFilesReportJson.class);
+        }
+
+        Assert.assertNotNull(metaDataReport);
+        Assert.assertNotNull(methodIdsReport);
+        Assert.assertNotNull(methodReferencesReport);
+        Assert.assertNotNull(brokenFilesReport);
+
+        Assert.assertEquals(metaDataReport.getHostName(), "hostName");
+        Assert.assertEquals(metaDataReport.getAnalyzedDirectory(), "analyzedDirectory");
+        Assert.assertEquals(metaDataReport.getCodeLocationName(), "codeLocationName");
+
+        Map<String, ReferencedMethodUsesJson> usesById = methodReferencesReport.getMethodUses().stream()
+                .collect(Collectors.toMap(use -> use.getMethod().getId(), Functions.identity()));
+
+        Set<String> signatureIds = methodIdsReport.getMethodIds().stream()
+                .map(MethodIdJson::getSignature)
+                .collect(Collectors.toSet());
+
+        Assert.assertEquals(signatureIds.size(), usesById.size());
+        Assert.assertTrue(signatureIds.containsAll(usesById.keySet()));
+
+        ReferencedMethodUsesJson resultJson = usesById.get("n0VJzGLYjELWet+V5A3IxQgkG/kQRR3P0OTmEpiScZs=");
+
+        Assert.assertNotNull(resultJson, "Expected reference ID not found, check that ID generation logic in ReportGenerator was not changed unintentionally.");
+        Assert.assertEquals(resultJson.getMethod().getId(), "n0VJzGLYjELWet+V5A3IxQgkG/kQRR3P0OTmEpiScZs=",
+                "Referenced Method ID was unexpected value, check that ID generation logic in ReportGenerator was not changed unintentionally.");
+        Assert.assertEquals(resultJson.getMethod().getMethodName(), "methodName");
+        Assert.assertEquals(resultJson.getMethod().getMethodOwner(), "methodOwner");
+        Assert.assertEquals(resultJson.getMethod().getOutput(), "output");
+        Assert.assertEquals(resultJson.getMethod().getInputs(), Collections.singletonList("input"));
+        Assert.assertEquals(resultJson.getUses().size(), 1);
+        Assert.assertTrue(resultJson.getUses().contains(new MethodUseJson("use", 1)));
+
+        ReferencedMethodUsesJson resultJsonNullLineNumber = usesById.get("lnTAAs55NfVEpFXFtKvJ5KmvK6Z6En4jHzP2Xp3PL7g=");
+
+        Assert.assertNotNull(resultJsonNullLineNumber,
+                "Expected reference ID not found, check that ID generation logic in ReportGenerator was not changed unintentionally.");
+        Assert.assertEquals(resultJsonNullLineNumber.getMethod().getId(), "lnTAAs55NfVEpFXFtKvJ5KmvK6Z6En4jHzP2Xp3PL7g=",
+                "Referenced Method ID was unexpected value, check that ID generation logic in ReportGenerator was not changed unintentionally.");
+        Assert.assertEquals(resultJsonNullLineNumber.getMethod().getMethodName(), "methodName2");
+        Assert.assertEquals(resultJsonNullLineNumber.getMethod().getMethodOwner(), "methodOwner2");
+        Assert.assertEquals(resultJsonNullLineNumber.getMethod().getOutput(), "output2");
+        Assert.assertEquals(resultJsonNullLineNumber.getMethod().getInputs(), Collections.singletonList("input2"));
+        Assert.assertEquals(resultJsonNullLineNumber.getUses().size(), 1);
+        Assert.assertTrue(resultJsonNullLineNumber.getUses().contains(new MethodUseJson("use", null)));
+
+        Assert.assertNotNull(brokenFilesReport.getBrokenFiles());
+        Assert.assertEquals(brokenFilesReport.getBrokenFiles().size(), 1);
+
+        BrokenFileJson brokenFileJson = brokenFilesReport.getBrokenFiles().get(0);
+        Assert.assertNotNull(brokenFileJson);
+        Assert.assertEquals(brokenFileJson.getPath(), "broken/BFile.class");
+        Assert.assertEquals(brokenFileJson.getError(), "Caused by TestException at MethodUseAnalyzer.java:123");
+    }
+
+    @Test
     public void partitionedReport() throws Exception {
         Multimap<ReferencedMethod, MethodUse> references = HashMultimap.create();
+        Map<Path, String> brokenFiles = new HashMap<>();
 
         for (int i = 0; i < 2000; i++) {
             references.put(new ReferencedMethod("methodOwner" + i, "methodName" + i, Collections.singletonList("input" + i), "output" + i),
                     new MethodUse("use" + i, i));
+            brokenFiles.put(Paths.get("broken/BFile" + i + ".class"), "Caused by TestException at MethodUseAnalyzer.java:123 " + i);
         }
 
-        Path result = reportGenerator.generateReport(references, testReportDirectory, "simpleReport");
+        Path result = reportGenerator.generateReport(references, brokenFiles, testReportDirectory, "simpleReport");
 
         Assert.assertNotNull(result);
         Assert.assertTrue(Files.exists(result));
@@ -167,19 +273,25 @@ public class ReportGeneratorTest {
         Path expectedMetaData = resultExpandedDirectory.resolve("metaData.json");
         Path expectedReferencedMethod1 = resultExpandedDirectory.resolve("referenced-methods").resolve("referenced-methods-0.json");
         Path expectedReferencedMethodUses1 = resultExpandedDirectory.resolve("referenced-method-uses").resolve("referenced-method-uses-0.json");
+        Path expectedBrokenFiles1 = resultExpandedDirectory.resolve("broken-files").resolve("broken-files-0.json");
         Path expectedReferencedMethod2 = resultExpandedDirectory.resolve("referenced-methods").resolve("referenced-methods-1.json");
         Path expectedReferencedMethodUses2 = resultExpandedDirectory.resolve("referenced-method-uses").resolve("referenced-method-uses-1.json");
+        Path expectedBrokenFiles2 = resultExpandedDirectory.resolve("broken-files").resolve("broken-files-1.json");
 
         Assert.assertTrue(Files.exists(expectedReferencedMethod1), "Expected " + expectedReferencedMethod1 + " to exist");
         Assert.assertTrue(Files.exists(expectedReferencedMethodUses1), "Expected " + expectedReferencedMethodUses1 + " to exist");
+        Assert.assertTrue(Files.exists(expectedBrokenFiles1), "Expected " + expectedBrokenFiles1 + " to exist");
         Assert.assertTrue(Files.exists(expectedReferencedMethod2), "Expected " + expectedReferencedMethod2 + " to exist");
         Assert.assertTrue(Files.exists(expectedReferencedMethodUses2), "Expected " + expectedReferencedMethodUses2 + " to exist");
+        Assert.assertTrue(Files.exists(expectedBrokenFiles2), "Expected " + expectedBrokenFiles2 + " to exist");
 
         MetaDataReportJson metaDataReport = null;
         MethodIdsReportJson methodIdsReport1 = null;
         MethodReferencesReportJson methodReferencesReport1 = null;
+        BrokenFilesReportJson brokenFilesReport1 = null;
         MethodIdsReportJson methodIdsReport2 = null;
         MethodReferencesReportJson methodReferencesReport2 = null;
+        BrokenFilesReportJson brokenFilesReport2 = null;
 
         try (BufferedReader reader = Files.newBufferedReader(expectedMetaData)) {
             metaDataReport = GSON.fromJson(reader, MetaDataReportJson.class);
@@ -201,11 +313,21 @@ public class ReportGeneratorTest {
             methodReferencesReport2 = GSON.fromJson(reader, MethodReferencesReportJson.class);
         }
 
+        try (BufferedReader reader = Files.newBufferedReader(expectedBrokenFiles1)) {
+            brokenFilesReport1 = GSON.fromJson(reader, BrokenFilesReportJson.class);
+        }
+
+        try (BufferedReader reader = Files.newBufferedReader(expectedBrokenFiles2)) {
+            brokenFilesReport2 = GSON.fromJson(reader, BrokenFilesReportJson.class);
+        }
+
         Assert.assertNotNull(metaDataReport);
         Assert.assertNotNull(methodIdsReport1);
         Assert.assertNotNull(methodReferencesReport1);
+        Assert.assertNotNull(brokenFilesReport1);
         Assert.assertNotNull(methodIdsReport2);
         Assert.assertNotNull(methodReferencesReport2);
+        Assert.assertNotNull(brokenFilesReport2);
 
         Assert.assertEquals(metaDataReport.getHostName(), "hostName");
         Assert.assertEquals(metaDataReport.getAnalyzedDirectory(), "analyzedDirectory");
@@ -214,6 +336,8 @@ public class ReportGeneratorTest {
         // Check partitioning numbers are correct
         Assert.assertEquals(methodIdsReport1.getMethodIds().size(), 1000);
         Assert.assertEquals(methodIdsReport2.getMethodIds().size(), 1000);
+        Assert.assertEquals(brokenFilesReport1.getBrokenFiles().size(), 1000);
+        Assert.assertEquals(brokenFilesReport2.getBrokenFiles().size(), 1000);
 
         Set<String> allIds = Stream.concat(methodIdsReport1.getMethodIds().stream(), methodIdsReport2.getMethodIds().stream())
                 .map(MethodIdJson::getSignature)